+++ src/OszkApi/Startup.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using OszkConnector.Models;
+using OszkConnector.Repository;
 
 namespace OszkApi
 {
